Working catalyst app with basic authentication
Authentication is working but credentials are currently plain-text
passwords in the config. Need to switch to the DBIC plugin.

The login form is working but isn't styled. HTML::FormHandler is a
peculiarly perverse instrument of torture but it should be possible to
style the form with bootstrap classes with some faffing about.