Cleaned up resources layout and login form
There were still some remnants of the dancer app layout lying around, so
this commit removes those directories and replaces the BMV tarball in
the correct location in the catalyst app.

The login form was still being rendered using the custom
HTML::FormHandler form, which wasn't working particularly well, so this
commit also removes the custom form and switches to rendering it using
a template.