Release candidate catalyst-based app
This commit includes changes across various files to fix problems with
the catalyst app, and, on top of that, it includes a mechanism to
specify that an action should require authentication by login or HMAC
header. Includes a jiffy script to send HMAC authenticated requests, to
make testing that bit easier.