Merging selected changes from the release branch
The initial_catalyst_release branch got the updated checklist and
validator code, which needs to be brought back across to the master. The
various scripts and configs for running the site in production were also
updated and need to be merged into master.

Fixed a bug in the site CSS which was making the navbar the wrong size.

Added two new shell files for setting up the shell environment for
running the server during development (setup_dev_environment.sh) and
when running tests (setup_test_environment.sh).