Disable sign-ins with a configuration option
Added checks to the navbar and NeedsAuth role to prevent signing in when
the config options "disable_signin" is set to true. The main block to
signing in will be the disabled "sign in" button in the navbar, but the
role should also drop any scripted requests that aren't using the form.