Fix a vulnerability with uploaded file handling
There were two potential problems with the initial attempt at this, as
pointed out by bt5.

First and most obviously, the UUID was not being detained for requests
to download the validated file. Second, the uploaded file was being
stored with the filename that was provided by the user, opening up the
possibility of the app wandering around the file system at the user's
behest.

This commit fixes those two issues. The uploaded file is now stored in a
directory named with a UUID. The UUID is validated when it's sent back
in the request to download the validated file. The validated file is now
stored with a generic name, which is not supplied by the user. The
original filename is stored in a metadata file next to the actual file,
and is only used to generate a Content-Disposition header when handing
back the file itself.