Add mechanism for removing old uploaded files
The validation tool will gradually collect uploaded files, which need to
be removed periodically. Adding a cron job is tricky because the server
runs as a restricted user, so the cron job would also need to run as
that user. To avoid the administrative hassles of setting up that cron
job, this commit adds a private action that will remove cached uploaded
files older than a configured age.

The action can be triggered by an external request, courtesy of
Catalyst::Plugin::Scheduler. The request must come from the server
itself, but it can be generated by a curl command in a crontab, and that
cron entry can be added by a non-restricted user.