Switch to using caching to store validated files
Writing files to disk will only work if user sessions can be locked to
one host, otherwise a user could find themselves uploading and
validating their file on one node, but trying to download it on the
other. Caching with a shared cache (using memcached in production)
should avoid that problem.

This commit throws away the code that saves the uploaded file to disk,
and the Catalyst::Plugin::Scheduler event method, which are no longer
needed...