Fix bug causing missing favicons
Added "build" blocks for images used as favicons.

There's a bug (or strange feature) in grunt.usemin which causes it not
to modify revisioned filenames in meta tags where the filename is
specified with a content attribute. That's how favicon "tiles" are
specified for MS devices, so the favicon tiles were being omitted
for those clients.

It's fixable in grunt.usemin code but it's easier to
add a workaround: use "href" rather than "content" in the tags and
usemin correctly versions the image files...