Remove certain links if signins are disabled
In the privacy page we have links to the forms that let the user change
their password or reset their API key. These won't work unless the
server is configured to accept signins.

In order to avoid the user seeing a crude error message when they click
on those links while signins are disabled, this commit tweaks the
template so that it only adds the links if signins are enabled.