Store API key as hashed digest; simplify Authorization header
We should handle the API key in the same was as the user password,
storing it as a hashed digest. Doing that means we can't use the simple
HMAC authentication scheme that we were using previously: since we don't
have the plain-text of the API key, we can't calculate the HMAC on the
server.

The change to storing hashed digests in the DB was implemented in
Bio::HICF::Schema in commit
https://github.com/sanger-pathogens/Bio-HICF-Schema/commit/b7fbb6875583c15d2985065eb2cbc3b252348057

Now, since we're doing everything over SSL, we can use a simplified
scheme in which the user sends their username and API key in the
Authorization header and we check the hash against the database using
a call to the user object.