Fix bug showing catalyst "come back later" response
If the validation process failed within one of the two try/catch blocks,
the site would end up showing the default catalyst error page embedded
in the main validation page. We can't simply "return" from within a
catch block, so this fix catches the exception and sets a flag, which
then triggers a real "return" from outside the catch block.