Update tests to handle split databases
The switch to two databases breaks a lot of tests, since they relied on
Test::DBIx::Class to set up the single, unified database that catalyst
read. Now that we have two DBs, one rw, one ro, we need to change the
way the tests work.

This commit updates the test scripts to use two test SQLite databases.
The paths to the two DB files are set in the config files. The old
"fixtures" aren't needed any longer, since the same data is simply
pre-loaded in the DB files.