Add catch for when content-type param is invalid
If the content-type param is added to the URL and is valid, catalyst
will try to return the response using that content type. If the content
type isn't valid, e.g. "content-type=application/jso", it falls back to
treating the response as if it comes from a browser.

Previously there was a problem with the handling of that case, because
there was no template specified and catalyst was looking for one in the
default location, where it will never find one.

Fix the problem by specifying a template earlier in the request life
cycle, so that it's available even for valid REST responses.