Add functionality to support DataTables
The DataTables table module can do paging, sorting and filtering on the
client side, but loading large tables is inherently slow, so it can also
get data from the server and allow the server to handle pagination,
sorting and filtering instead. This commit adds the necessary
functionality to the "/sample" action.