Refactor methods that paginate, filter and sort
Tidy up and rationalise the methods that slice-and-dice the resultset
that's handed into "_get_sample_data", to make it applicable to various
starting sets. The idea is to make that method work with a ResultSet
containing all samples, or a subset of samples such as those from a
given organism.