Add REST and HTML endpoints for sample data
Add chained actions to handle various types of request for sample data.

This controller can now return all samples, either via a page that
embeds it in an AJAX-loaded table or via REST as JSON, YAML, CSV, etc
(/samples). It can also return a subset of all samples
(/samples/<start>/<length>), all samples for a given organism
(/samples/<organism>) and a subset of samples for an organism
(/samples/<organism>/<start>/<length>).

It's all a bit convoluted and messy but the functionality is there.