Fix sorting by re-ordering table columns
The previous fix for broken sorting was to correct for the additional
AMR column that was inserted at the start of the table. That causes
problems because, without adding a hacky work-around for the very first
column, sample_id, you can't sort on sample_id.

This commit gets around the problem by simply re-ordering the columns
and putting the AMR back at the end of the tables, so it doesn't
interfere with the column indices that sorting needs. It's not ideal but
it's the least worst option right now.