Save table state for 5 min
DataTables can save the state of the table in various ways, which allows
niceties like retaining sort order or page number when going forwards to
a sample page and then back to the table of all samples.

Unfortunately, it retains that state even when going from a table of all
samples to another table that's filtered by, say, organism name. This
means that the new table can appear on a page other than 1 and possibly
with a different sort order to the expected one.

State-saving is useful but we'll limit it to 5 minutes, rather than per
browser session.