Show "n/a" for unknowns; catch logged out sessions
Add a default cell renderer for the samples table, which replaces empty
cell values with "n/a" and greys out "unknown" values.

If a user let their sign in session expire and then tries to load a new
page in the samples table, they would previously get an unhelpful error
message from DataTables. Add a trap for a "401 Unauthorized" responses
to the ajax call, so that we can show a slightly more helpful message
and redirect the user to the login page.