Make table draggable; fix event handling
Make the table a jQuery UI draggable, so that it can be scrolled with
the mouse wheel or dragged.

Fixed the calls that add listeners. They were using the wrong name for
the target div, so listeners weren't being added consistently.

Improved the positioning and wording of the links. Made the text more
informative, hopefully.