Fix sort links in summary text
There are two small DataTable tables in the page. We have links next to
one of the, which should let the user sort the table in different ways.
There was a bug in the JS that hooks in those links, meaning that they
didn't point to the table object and therefore didn't work. This commit
fixes that.