Revert to a plain form for sign in/out button
The sign in form was previously handled by a JS object, but that's just
a hang over from when it might have triggered a pop up sign in form. Now
that it just goes straight to the login page, we don't need and JS
involved and we can make it a plain form.

This commit gets rid of the "login" js object and makes the form a
vanilla web form. It also tidies up the markup and fixes the CSS, which
is necessary because of the changes in the DOM structure.