Tidy up for use with front-end tests
Use the two test databases directly, by copying them from the perl test
data directory, rather than trying to instantiate from an SQL script.

Move the test server to port 3001 to differentiate it from the
development server, which is likely to be pointing at a MySQL database
rather than the little SQLite DBs that we need for the tests.