Fix problem with page layout and ToC
There was a problem with the positioning of the table of contents when
the window was landscape rather than portrait, noticeable when the
window was resized. The ToC would overlap the page contents and resize
strangely itself.

Fixed by tweaking the HTML and CSS to make the margins and widths behave
better.