Compress the validated file before caching
When a file has been validated in the online validation tool, we cache
the validated file, rather than storing it on disk. The default cache is
memcached.

There was a problem with large CSV files, especially after adding a
large number of error messages, in that the file size exceeded the
default maximum item size for memcached. We should definitely increase
that limit but in the meantime it also makes sense to compress the stuff
that we cache.

This commit adds a few lines that compress the validated file before
caching it and uncompress it when retrieving it. The compression ratio
for CSVs is pretty impressive, so that will buy us a lot of overhead and
shouldn't impact performance greatly.