Add extra hidden column to table
The location field is converted to a location description in the
controller, but when downloading or saving the table contents, it might
be important to have the original GAZ term. This commit adds the GAZ
term as a hidden column in the samples table and makes it visible in the
JSON or CSV rendering of the table that is saved by the buttons.