Fix filtering to make it work when saving
The data generated by the "save CSV" and "save JSON" buttons previously
wasn't filtered in the same way as the table contents. This commit
re-works the way that filtering is done, so that it copes with multiple
filter terms (from a click in the table, such as manifest_id, and from
the table search field).