Handle undefined fields when building samples table
When DataTables builds the HTML for the samples table, it uses a
renderer function that tweaks or formats the raw values from the server
before putting them into the table. Previously, if there was an
undefined value in the hash that came from the server (e.g.
location_description), the "switch" statement in the renderer would try
to put the value (undef) into the table directly, causing an alert from
DataTables.

This commit just adds a "case undefined", which makes the renderer treat
undefined values the same as empty strings, i.e. it replaces them by
"n/a" in the output.